First working WZM format implementation [Be warned - bugs territory!]
WZM coexists with PIE inside a single structure, which is, due to public nature, is more like a timebomb currently, because not all data is filled for PIE usage.
So use WZM objects in whitelisted mode for now: weapons, props and usual structures can be coped with for tests.
Don't touch models which are animated in any way (mesh or texture), effects (hardcoded flags aren't ported yet), defense structures (nasty public points usage)...